Refactor MetadataStore methods
